requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
discord.py==1.7.3
docopt==0.6.2
idna==3.3
multidict==5.2.0
pipdeptree==2.2.0
pipreqs==0.4.11
requests==2.26.0
typing_extensions==4.0.1
urllib3==1.26.7
yarg==0.1.9
yarl==1.7.2
